pyproject.toml
[tool.poetry]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = ["Mario Toribio <yuniormtr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
langchain = "^0.3.7"
langchain-openai = "^0.2.6"
langgraph = "^0.2.45"
langchain-ollama = "^0.2.0"
grandalf = "^0.8"
python-decouple = "^3.8"
langchain-groq = "^0.2.1"
langchain-community = "^0.3.5"
duckduckgo-search = "^6.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
